==> Preprocessing/3_Prepare_visualisation_for_cellxgene.Rmd <==
---
title: "Batch correction across tissues and patients"
author: "Karol Nowicki-Osuch"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float:
      toc_collapsed: true
    toc_depth: 3
    number_sections: false
    theme: paper
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
# Rscript -e "library(rmarkdown); render('Rebuttal_All.Rmd')"
knitr::opts_chunk$set(echo = TRUE)
```

# Read in the data

```{r libraries, message=FALSE}
library(scran)
library(anndata)
library(zellkonverter)
```

# Introduce global cluster names after removal of doublets

After the analysis performed in the previous steps of the preprocessing pipeline, prepare data for the visualisation using cellxgene tool. The data will be published on the cell atlas website. https://cellxgene.cziscience.com/

To visualise, run:
cellxgene launch ~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells.h5ad --host kno01.cancerdynamics.columbia.edu --disable-diffexp --disable-annotations -p 5006


```{r}
#Read cell types
sce.visualize<-readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_after_exclusion2.rds")
colData(sce.visualize) <- colData(sce.visualize)[,colnames(colData(sce.visualize)) %in% c("Sample",
                                                                                          "Barcode",
                                                                                          "Patient",
                                                                                          "Tissue",
                                                                                          "Batch",
                                                                                          "Sample_Barcode",
                                                                                          "sum",
                                                                                          "detected",
                                                                                          "Study",
                                                                                          "Patient_type",
                                                                                          "Patient_status",
                                                                                          "MT.prop",
                                                                                          "sizeFactor",
                                                                                          "Global_cluster_selected",
                                                                                          "Celltypes_global",
                                                                                          "Tissuetypes_global")]
sce.visualize$Detailed_Cell_Type <- sce.visualize$Celltypes_global
tracking <- paste(sce.visualize$Sample, sce.visualize$Barcode)

# Intruduce individual cell types
## Endothelial cells
sce.tmp <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_Endothelial.rds")
tracking2 <- paste(sce.tmp$Sample, sce.tmp$Barcode)
sce.visualize$Detailed_Cell_Type[match(tracking2, tracking)] <- sce.tmp$Celltypes_Endothelial

## Fibroblasts
sce.tmp <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_Fibroblasts.rds")
tracking2 <- paste(sce.tmp$Sample, sce.tmp$Barcode)
sce.visualize$Detailed_Cell_Type[match(tracking2, tracking)] <- sce.tmp$Fibroblast_Sarah

## Myeloid
sce.tmp <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_Myeloid.rds")
tracking2 <- paste(sce.tmp$Sample, sce.tmp$Barcode)
sce.visualize$Detailed_Cell_Type[match(tracking2, tracking)] <- sce.tmp$Celltypes_MyeloidCells

## T Cells
sce.tmp <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_T-Cells_new_correction.rds")
tracking2 <- paste(sce.tmp$Sample, sce.tmp$Barcode)
sce.visualize$Detailed_Cell_Type[match(tracking2, tracking)] <- sce.tmp$Celltypes_TCells
sce.visualize$Celltypes_global[match(tracking2, tracking)] <- "T-Cells"

## B Cells
sce.tmp <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_PlasmaandT-Cells.rds")
tracking2 <- paste(sce.tmp$Sample, sce.tmp$Barcode)
sce.visualize$Detailed_Cell_Type[match(tracking2, tracking)] <- sce.tmp$Celltypes_BCells

write.table(unique(colData(sce.visualize)[,"Detailed_Cell_Type", drop = FALSE]), file = "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/Cell_types.txt", quote = FALSE, sep = "\t", row.names = FALSE, col.names = TRUE)
cell.types <- read.delim( file = "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/Cell_types_annotated.txt")


sce.visualize$cell_type_ontology_term_id <- cell.types$cell_type_ontology_term_id[match(sce.visualize$Detailed_Cell_Type, cell.types$Detailed_Cell_Type)]

write.table(unique(colData(sce.visualize)[,"Sample", drop = F]), file = "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/Patient_Data.txt", quote = FALSE, sep = "\t", row.names = FALSE, col.names = TRUE)


studies <- c("Fitz" = "Nowicki-Osuch et al. 2022",
             "Ji" = "Sathe et al. 2020",
             "Li" = "Zhang et al. 2019",
             "Wang" = "Wang et. al. 2020"
)

sce.visualize$Study <- studies[sce.visualize$Study]

sample.data <- read.delim( file = "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/Patient_Data_for_visualisation.txt")
sample.data <- sample.data[match(sce.visualize$Sample,sample.data$Sample),!colnames(sample.data) %in% colnames(colData(sce.visualize))]
colData(sce.visualize) <- cbind(colData(sce.visualize), sample.data)

metadata(sce.visualize) <- list(schema_version = "2.0.0", title = "UMAP of all data", X_normalization = "Log2(count +1). Data was scaled using size factors. Packages: batchelor and scran")

sce.visualize$donor_id <- sce.visualize$Patient

for ( 
  Column in c("Sample",
              # "Patient",
              # "Tissue",
              "Batch",
              "Sample_Barcode",
              "Study",
              "Patient_type",
              "Patient_status",
              "Global_cluster_selected",
              "Celltypes_global",
              "Tissuetypes_global",
              "Detailed_Cell_Type",
              "cell_type_ontology_term_id",
              # "Assay", 
              "assay_ontology_term_id",
              "tissue_ontology_term_id",
              "disease_ontology_term_id",
              # "Gender",
              "sex_ontology_term_id",
              "organism_ontology_term_id",
              "ethnicity_ontology_term_id",
              "development_stage_ontology_term_id",
              "donor_id")
) {
  colData(sce.visualize)[,Column] <- as.factor(colData(sce.visualize)[,Column])
}

rowData(sce.visualize) <- rowData(sce.visualize)[,1:2]

saveRDS(sce.visualize, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_withBarcode_data.rds")

sce.visualize <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_withBarcode_data.rds")

colnames(sce.visualize) <- paste(sce.visualize$Sample, sce.visualize$Barcode, sep = "-")

colData(sce.visualize) <- colData(sce.visualize)[,!names(colData(sce.visualize)) %in% c("Barcode",
                                                                                        "Patient",
                                                                                        # "Tissue",
                                                                                        "Assay", 
                                                                                        "Gender")]
name.changes <-c("NE" = "NE",
                 "NSCJ" = "NSCJ",
                 "SMG" = "SMG",
                 "GM" = "E-GM",
                 "BE" = "BE-IM",
                 "BSCJ" = "BSCJ",
                 "NG" = "NGC",
                 "NGB" = "NGB",
                 "NAG" = "NAG",
                 "CAG" = "CAG",
                 "GIM" = "GIM",
                 "CIM" = "CIM",
                 "ND" = "ND",
                 "Ileum" = "Ileum", 
                 "Colon" = "Colon",
                 "Rectum" = "Rectum")

sce.visualize$Tissue <- name.changes[sce.visualize$Tissue]

adata <- zellkonverter::SCE2AnnData(sce.visualize, X_name = "logcounts")
# adata$raw <- anndata::AnnData(counts(sce.visualize))
names(adata$obsm)[4] <- paste0(".",names(adata$obsm)[4])
names(adata$obsm)  <- paste0("X_",names(adata$obsm))
# adata$obsm <- adata$obsm[c(4,1:3,5:7)]

write_h5ad(adata, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells.h5ad", compression = "gzip")

saveRDS(sce.visualize, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells.rds")


```

## Columnar cells

```{r}

sce.visualize<-readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_withBarcode_data.rds")

colnames(sce.visualize) <- paste(sce.visualize$Sample, sce.visualize$Barcode, sep = "-")

sce.specific <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_Columnar.rds")

sce.visualize.specific <- sce.visualize[,match( paste(sce.specific$Sample, sce.specific$Barcode), paste(sce.visualize$Sample, sce.visualize$Barcode))]
reducedDims(sce.visualize.specific) <- reducedDims(sce.specific)
colData(sce.visualize.specific) <- cbind(colData(sce.visualize.specific), colData(sce.specific)[,c("Columnar_clusters"), drop = FALSE])
sce.visualize.specific$Columnar_clusters <- as.factor(sce.visualize.specific$Columnar_clusters)
metadata(sce.visualize.specific)$title <- "UMAP of Columnar cells"
sce.visualize.specific$Tissue <- name.changes[sce.visualize.specific$Tissue]

saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Columnar_Cells_withBarcode_data.rds")
# sce.visualize.specific <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Columnar_Cells_withBarcode_data.rds")

colData(sce.visualize.specific) <- colData(sce.visualize.specific)[,!names(colData(sce.visualize.specific)) %in% c("Barcode",
                                                                                        "Patient",
                                                                                        # "Tissue",
                                                                                        "Assay",
                                                                                        "Gender")]

adata <- zellkonverter::SCE2AnnData(sce.visualize.specific, X_name = "logcounts")
# adata$raw <- anndata::AnnData(counts(sce.visualize))
names(adata$obsm)[4] <- paste0(".",names(adata$obsm)[4])
names(adata$obsm)  <- paste0("X_",names(adata$obsm))
# adata$obsm <- adata$obsm[c(4,1:3,5:7)]

write_h5ad(adata, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Columnar_Cells.h5ad", compression = "gzip")

saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Columnar_Cells.rds")

```


## Endothelial cells

```{r}

# sce.visualize<-readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_withBarcode_data.rds")
colnames(sce.visualize) <- paste(sce.visualize$Sample, sce.visualize$Barcode, sep = "-")

sce.specific <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_Endothelial.rds")

sce.visualize.specific <- sce.visualize[,match( paste(sce.specific$Sample, sce.specific$Barcode), paste(sce.visualize$Sample, sce.visualize$Barcode))]
reducedDims(sce.visualize.specific) <- reducedDims(sce.specific)
metadata(sce.visualize.specific)$title <- "UMAP of Endothelial cells"
sce.visualize.specific$Tissue <- name.changes[sce.visualize.specific$Tissue]
saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Endothelial_Cells_withBarcode_data.rds")

colData(sce.visualize.specific) <- colData(sce.visualize.specific)[,!names(colData(sce.visualize.specific)) %in% c("Barcode",
                                                                                        "Patient",
                                                                                        # "Tissue",
                                                                                        "Assay",
                                                                                        "Gender")]

adata <- zellkonverter::SCE2AnnData(sce.visualize.specific, X_name = "logcounts")
# adata$raw <- anndata::AnnData(counts(sce.visualize))
names(adata$obsm)[4] <- paste0(".",names(adata$obsm)[4])
names(adata$obsm)  <- paste0("X_",names(adata$obsm))
# adata$obsm <- adata$obsm[c(4,1:3,5:7)]

write_h5ad(adata, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Endothelial_Cells.h5ad", compression = "gzip")

saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Endothelial_Cells.rds")

```

## Fibroblasts cells

```{r}

# sce.visualize<-readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_withBarcode_data.rds")
colnames(sce.visualize) <- paste(sce.visualize$Sample, sce.visualize$Barcode, sep = "-")

sce.specific <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_Fibroblasts.rds")

sce.visualize.specific <- sce.visualize[,match( paste(sce.specific$Sample, sce.specific$Barcode), paste(sce.visualize$Sample, sce.visualize$Barcode))]
reducedDims(sce.visualize.specific) <- reducedDims(sce.specific)
metadata(sce.visualize.specific)$title <- "UMAP of Fibroblasts cells"
sce.visualize.specific$Tissue <- name.changes[sce.visualize.specific$Tissue]
saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Fibroblasts_Cells_withBarcode_data.rds")

colData(sce.visualize.specific) <- colData(sce.visualize.specific)[,!names(colData(sce.visualize.specific)) %in% c("Barcode",
                                                                                        "Patient",
                                                                                        # "Tissue",
                                                                                        "Assay",
                                                                                        "Gender")]

adata <- zellkonverter::SCE2AnnData(sce.visualize.specific, X_name = "logcounts")
# adata$raw <- anndata::AnnData(counts(sce.visualize))
names(adata$obsm)[4] <- paste0(".",names(adata$obsm)[4])
names(adata$obsm)  <- paste0("X_",names(adata$obsm))
# adata$obsm <- adata$obsm[c(4,1:3,5:7)]

write_h5ad(adata, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Fibroblasts_Cells.h5ad", compression = "gzip")

saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Fibroblasts_Cells.rds")

```


## Myeloid cells

```{r}

# sce.visualize<-readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_withBarcode_data.rds")
colnames(sce.visualize) <- paste(sce.visualize$Sample, sce.visualize$Barcode, sep = "-")

sce.specific <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_Myeloid.rds")

sce.visualize.specific <- sce.visualize[,match( paste(sce.specific$Sample, sce.specific$Barcode), paste(sce.visualize$Sample, sce.visualize$Barcode))]
reducedDims(sce.visualize.specific) <- reducedDims(sce.specific)
metadata(sce.visualize.specific)$title <- "UMAP of Myeloid cells"
sce.visualize.specific$Tissue <- name.changes[sce.visualize.specific$Tissue]
saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Myeloid_Cells_withBarcode_data.rds")

colData(sce.visualize.specific) <- colData(sce.visualize.specific)[,!names(colData(sce.visualize.specific)) %in% c("Barcode",
                                                                                        "Patient",
                                                                                        # "Tissue",
                                                                                        "Assay",
                                                                                        "Gender")]

adata <- zellkonverter::SCE2AnnData(sce.visualize.specific, X_name = "logcounts")
# adata$raw <- anndata::AnnData(counts(sce.visualize))
names(adata$obsm)[4] <- paste0(".",names(adata$obsm)[4])
names(adata$obsm)  <- paste0("X_",names(adata$obsm))
# adata$obsm <- adata$obsm[c(4,1:3,5:7)]

write_h5ad(adata, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Myeloid_Cells.h5ad", compression = "gzip")

saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/Myeloid_Cells.rds")

```

## Plasma and B-Cells cells

```{r}

# sce.visualize<-readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_withBarcode_data.rds")
colnames(sce.visualize) <- paste(sce.visualize$Sample, sce.visualize$Barcode, sep = "-")

sce.specific <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_PlasmaandT-Cells.rds")

sce.visualize.specific <- sce.visualize[,match( paste(sce.specific$Sample, sce.specific$Barcode), paste(sce.visualize$Sample, sce.visualize$Barcode))]
reducedDims(sce.visualize.specific) <- reducedDims(sce.specific)
metadata(sce.visualize.specific)$title <- "UMAP of Plasma and B-Cells cells"
sce.visualize.specific$Tissue <- name.changes[sce.visualize.specific$Tissue]
saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/PlasmaandB-Cells_Cells_withBarcode_data.rds")

colData(sce.visualize.specific) <- colData(sce.visualize.specific)[,!names(colData(sce.visualize.specific)) %in% c("Barcode",
                                                                                        "Patient",
                                                                                        # "Tissue",
                                                                                        "Assay",
                                                                                        "Gender")]

adata <- zellkonverter::SCE2AnnData(sce.visualize.specific, X_name = "logcounts")
# adata$raw <- anndata::AnnData(counts(sce.visualize))
names(adata$obsm)[4] <- paste0(".",names(adata$obsm)[4])
names(adata$obsm)  <- paste0("X_",names(adata$obsm))
# adata$obsm <- adata$obsm[c(4,1:3,5:7)]

write_h5ad(adata, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/PlasmaandB-Cells_Cells.h5ad", compression = "gzip")

saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/PlasmaandB-Cells_Cells.rds")

```

## T-Cells cells

```{r}

# sce.visualize<-readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_withBarcode_data.rds")
colnames(sce.visualize) <- paste(sce.visualize$Sample, sce.visualize$Barcode, sep = "-")

sce.specific <- readRDS("~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_corrected_T-Cells_new_correction.rds")

sce.visualize.specific <- sce.visualize[,match( paste(sce.specific$Sample, sce.specific$Barcode), paste(sce.visualize$Sample, sce.visualize$Barcode))]
reducedDims(sce.visualize.specific) <- reducedDims(sce.specific)
metadata(sce.visualize.specific)$title <- "UMAP of T-Cells cells"
sce.visualize.specific$Tissue <- name.changes[sce.visualize.specific$Tissue]
saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/T-Cells_Cells_withBarcode_data.rds")

colData(sce.visualize.specific) <- colData(sce.visualize.specific)[,!names(colData(sce.visualize.specific)) %in% c("Barcode",
                                                                                        "Patient",
                                                                                        # "Tissue",
                                                                                        "Assay",
                                                                                        "Gender")]

adata <- zellkonverter::SCE2AnnData(sce.visualize.specific, X_name = "logcounts")
# adata$raw <- anndata::AnnData(counts(sce.visualize))
names(adata$obsm)[4] <- paste0(".",names(adata$obsm)[4])
names(adata$obsm)  <- paste0("X_",names(adata$obsm))
# adata$obsm <- adata$obsm[c(4,1:3,5:7)]

write_h5ad(adata, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/T-Cells_Cells.h5ad", compression = "gzip")

saveRDS(sce.visualize.specific, "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/T-Cells_Cells.rds")

```

# Cancer Data



```{r}
name.changes <-c("NE" = "NE",
                 "NSCJ" = "NSCJ",
                 "SMG" = "SMG",
                 "GM" = "E-GM",
                 "BE" = "BE-IM",
                 "BSCJ" = "BSCJ",
                 "NG" = "NGC",
                 "NGB" = "NGB",
                 "NAG" = "NAG",
                 "CAG" = "CAG",
                 "GIM" = "GIM",
                 "CIM" = "CIM",
                 "ND" = "ND",
                 "Ileum" = "Ileum", 
                 "Colon" = "Colon",
                 "Rectum" = "Rectum",
                 "Gastric_Tumor" = "Gastric_Tumor",
                 "Normal_Adjacent" = "Normal_Adjacent")
sce.visualize <- readRDS( "~/Dropbox/Postdoc/2021-05-19_Gastric_IM/Final/All_sce_selected_with_Cancer_corrected_labels.rds")
# sce.visualize <- sce
rowData(sce.visualize) <- rowData(sce.visualize)[,1:2]
sce.visualize$Global_cluster <- as.factor(sce.visualize$Global_cluster )
sce.visualize$Celltypes_global <- as.factor(sce.visualize$Celltypes_global)
#randomise data
set.seed(50014)
jumbled<-sample(ncol(sce.visualize), replace = FALSE)
sce.visualize<-sce.visualize[,jumbled]
sce.visualize$Tissue <- name.changes[sce.visualize$Tissue]
colData(sce.visualize) <- colData(sce.visualize)[,!names(colData(sce.visualize)) %in% c("Barcode",
                                                                                        "Patient",
                                                                                        # "Tissue",
                                                                                        "Assay",
                                                                                        "Gender")]
sce.visualize$Study[sce.visualize$Study == "Kumar at al. 2022"] <- "Kumar et al. 2022" 
metadata(sce.visualize)$title <- "UMAP of Cancer Data integration"
adata <- zellkonverter::SCE2AnnData(sce.visualize, X_name = "logcounts")
# adata$raw <- anndata::AnnData(counts(sce.visualize))
names(adata$obsm)[4] <- paste0(".",names(adata$obsm)[4])
names(adata$obsm)  <- paste0("X_",names(adata$obsm))
# adata$obsm <- adata$obsm[c(4,1:3,5:7)]

write_h5ad(adata, "/home/karolno/Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_with_Cancer_with_Cells.h5ad", compression = "gzip")

saveRDS(sce.visualize, "/home/karolno//Dropbox/Postdoc/2021-05-19_Gastric_IM/Visualisation/All_Cells_with_Cancer_with_Cells.rds")




```




# End Matter

To finish get session info:

```{r Endnote, include=TRUE, echo=TRUE, warning=FALSE, eval=TRUE, message=FALSE, tidy=TRUE}
sessionInfo()
```
